--- Tests/test_OsUtility.py
@@ -1,16 +1,0 @@
-from unittest import TestCase
-import Utility.OsUtility as OsUtility
-
-
-class TestOsUtility(TestCase):
-    def test_get_file_list(self):
-        self.fail("Not Implemented Yet")
-
-    def test_clean_up_folder(self):
-        self.fail("Not Implemented Yet")
-
-    def test_create_folder_if_not_exist(self):
-        self.fail("Not Implemented Yet")
-
-    def test_get_filename(self):
-        self.fail("Not Implemented Yet")
